rename refinement_cc_in_one to refinement_in_one
