+++ example/stochastic_subgrid/stochastic_subgrid_driver.cpp
@@ -20,7 +20,7 @@
 #include "interface/metadata.hpp"
 #include "interface/update.hpp"
 #include "mesh/meshblock_pack.hpp"
-#include "mesh/refinement_cc_in_one.hpp"
+#include "mesh/refinement_in_one.hpp"
 #include "parthenon/driver.hpp"
 #include "refinement/refinement.hpp"
 #include "stochastic_subgrid_driver.hpp"
