+++ src/CMakeLists.txt
@@ -165,8 +165,8 @@ add_library(parthenon
   mesh/meshblock_tree.cpp
   mesh/meshblock_tree.hpp
   mesh/meshblock.cpp
-  mesh/refinement_cc_in_one.cpp
-  mesh/refinement_cc_in_one.hpp
+  mesh/refinement_in_one.cpp
+  mesh/refinement_in_one.hpp
 
   outputs/formatted_table.cpp
   outputs/history.cpp
