+++ src/mesh/refinement_in_one.cpp
@@ -22,7 +22,7 @@
 
 #include "kokkos_abstraction.hpp"
 #include "mesh/mesh_refinement_ops.hpp"
-#include "mesh/refinement_cc_in_one.hpp"
+#include "mesh/refinement_in_one.hpp"
 
 namespace parthenon {
 namespace cell_centered_refinement {
